pyproject.toml
[project]
name = "ai-poject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "google-genai>=1.53.0",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
